+++ src/objectdetection/detectors/tfjs-mobilenet-object-detector.ts
@@ -1,7 +1,7 @@
 import {ObjectDetectionDetectedObject} from '@dps/mycms-commons/dist/commons/model/objectdetection-model';
 import * as mobilenet from '@tensorflow-models/mobilenet';
-import {MobileNet, MobileNetAlpha, MobileNetVersion} from '@tensorflow-models/mobilenet';
-import {Tensor3D} from '@tensorflow/tfjs';
+import {MobileNet, MobileNetAlpha, MobileNetVersion, ModelConfig} from '@tensorflow-models/mobilenet';
+import {Tensor3D} from '@tensorflow/tfjs-core';
 import {AbstractObjectDetector, DetectorInputRequirement} from '../abstract-object-detector';
 import {DetectorResultUtils} from '../utils/detectorresult-utils';
 import {TensorUtils} from '../utils/tensor-utils';
@@ -12,7 +12,7 @@ export class TFJsMobilenetObjectDetector extends AbstractObjectDetector {
     private detector: MobileNet;
     private mobileNetVersion: MobileNetVersion = 1;
     private mobileNetAlpha: MobileNetAlpha = 1;
-    private BASEURL = 'https://storage.googleapis.com/tfhub-tfjs-modules/google/imagenet/';
+    private BASEURL = 'https://tfhub.dev/google/imagenet/';
     private assetsDir = 'assets/models/mobilenet/';
     private maxPredictionNumber = 1;
     private readonly rootDir: string;
@@ -44,7 +44,12 @@ export class TFJsMobilenetObjectDetector extends AbstractObjectDetector {
 
     initDetector(): Promise<boolean> {
         return new Promise<boolean>((resolve, reject) => {
-            return mobilenet.load(this.mobileNetVersion, this.mobileNetAlpha, this.getModelBasePath() + 'model.json').then(model => {
+            const modelConfig: ModelConfig = {
+                version: this.mobileNetVersion,
+                alpha: this.mobileNetAlpha,
+                modelUrl:  this.getModelBasePath() + 'model.json'
+            }
+            return mobilenet.load(modelConfig).then(model => {
                 this.detector = model;
 
                 return resolve(true);